Fix: ust-consumer: Use new flush events or populate packet API
Observed issue
==============

When taking multiple subsequent snapshots of a channel in overwrite
mode that is not receiving new events, after a number of snapshots
equal to the configured number of sub-buffers, the existing events are
cleared as a packet is forcibly closed each time a snapshot record is
issued.

Solution
========

lttng-ust has a new flush API that allows subsequent active flushes
to be done with forcibly closing the current packet. The new API
may populate a stand-alone packet that can be sent after the rest
of the stream contents to allow readers to infer that there are
no events between the last event and the time the flush was done.

This change is transparent to the user.

Known drawbacks
===============

This depends on the new events flush API introduced in the following
lttng-ust commit[3]:

    Add new flush: flush events or populate packet
    Change-Id: I5a9324354d8bb10e1dafcc3d455508924adf7481

References
==========
[1]: https://github.com/lttng/lttng-tools/commit/de3fb857f034c208c135a10a3cdec2dfe43fbda6
[2]: https://lists.lttng.org/pipermail/lttng-dev/2016-December/026805.html
[3]: https://review.lttng.org/c/lttng-ust/+/12533

Change-Id: I6a097611e359725c1928a12e55607ab0ef53de39
Depends-on: lttng-ust: I5a9324354d8bb10e1dafcc3d455508924adf7481
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>