Fix: kernel-consumer: Use new flush or populate packet API
Observed issue
==============

When taking multiple subsequent snapshots of a channel in overwrite
mode that is not receiving new events, after a number of snapshots
equal to the configured number of sub-buffers the existing events are
cleared as a packet is forcibly closed each time a snapshot record is issued.

Solution
========

lttng-modules adds a new flush operation which is transparent to the
user. This flush operation performs an active flush and, if necessary,
populates a stand-alone packet which may be output after the rest of
the stream's sub-buffers to allow readers to determine stream
intersections.

Known drawbacks
===============

This depends on the new events flush API introduced in the following
lttng-modules commit[3]:

Add new flush: flush events or populate packet
Change-Id: Ie929a8128d9a2bc6bb53e78363dba00e776aa163

If the version of lttng-modules loaded does not support the new flush
operation, then the consumer will fall back to using the previous flushes.

References
==========
[1]: https://github.com/lttng/lttng-tools/commit/f22dd89135f6b1749cd75735f04e3cb309da12c1
[2]: https://lists.lttng.org/pipermail/lttng-dev/2016-December/026805.html
[3]: https://review.lttng.org/c/lttng-modules/+/12883

Change-Id: I21337aa1223b0ab6962b51869d71ff068361480f
Depends-on: lttng-ust: I5a9324354d8bb10e1dafcc3d455508924adf7481
Depends-on: lttng-modules: Ie929a8128d9a2bc6bb53e78363dba00e776aa163
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>