Tests: Add path check to prevent incorrect redirection
The current implementation of `run_lttng_cmd` does not support
redirecting `stdout` and `stderr` to the same file. In the current
implementation:

    $TESTDIR/../src/bin/lttng/$LTTNG_BIN "$@" 1> "$stdout_dest" 2> "$stderr_dest"

the redirection of stderr would overwrite the redirection of stdout.

Add a check to bail out if the redirection paths are the same since the
test will not be as expected.

Change-Id: Ia8883f4978e4c0e534db1422a5aebf392403e723
Signed-off-by: Erica Bugden <ebugden@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>