Fix: relayd: incorrect use of lfht add_unique API
cds_lfht_add_unique is misused in relay_index_add_unique. The address
of the node is used instead of the address of the key for both the
key parameter and to compute the hash of the index entry's key.

This caused problems in an upcoming change that changes the layout
of the lttng_ht node since the key is no longer the first member
of the node.

Change-Id: Iec67f78b8ced4f4e520dc0dbebda8d30e283aa38
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>