Maintain recording channel configuration objects in ltt_session
Maintain the state of event rules in the ltt_session's channel
configurations. This is laying the groundwork to maintain triggers (and
ultimately, event-rules) associated with map channels.

liblttng-ctl now converts the lttng_event specifications into
event-rules which are used by the session daemon to maintain recording
channel configurations.

The objective is to share the synchronization code of
tracers (essentially, the description of enablers and channel
configuration) with the session configurations when adding support for
maps.

As of this commit, the event-rule-configurations are maintained
internally, but aren't use to synchronize the tracer configuration.

Change-Id: I6067ef31ca4b9cead55cf4c9dc2bac8dd8ca3400
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>