Bump required clang-format version to 16
Bump the required clang-format version from 14 to 16. For reference,
this is the version the Babeltrace project uses [1], and it is checked
by the LTTng CI [2], so I think it makes sense to go to that version.
More recent versions might not be readily available on the CI machines
(mostly running Debian 12 at the moment).

FWIW, I think that the code changes generated by this version change are
improvements.

[1] https://github.com/efficios/babeltrace/blob/d16ccfd174984c3d18f4f4427e4e438b1a64730e/tools/format-cpp.sh#L7
[2] https://ci.lttng.org/view/Babeltrace/job/babeltrace_master_lint/

Change-Id: I188f04928357ac7562b1a2dce3238a622189b04a
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>