Tests: Start sessiond in python test environments with `--no-kernel`
Observed issue
==============

One of the goals of the lttngtest Environment is to be able to produce
separate test environments for each run in order to eventually run
some tests in parallel.

Solution
========

If `lttng-sessiond` is run as root, by default it will attempt to load
the lttng-modules kernel modules.

The session daemon is now started with `--no-kernel` to disable the
kernel domain and prevent that instance from loading the kernel
modules. This behaviour can be changed for specific tests that
exercise kernel tracing.

Known drawbacks
===============

The `--no-kernel` start is less representative of the default usage of
the session daemon.

Change-Id: Iacb85d0a33d6981bd038d57bf629d61117f9344d
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>