lttng-relayd: Add `--pid-file` parameter
Intended to be used a mechanism to get the PID of the process when
starting with `-b` or `-d`.

`lttng-sessiond` writes to a well-known path inside `$LTTNG_HOME`
instead.

The exit for the lttng-relayd `main()` function is changed from
`exit()` to `return`. When using `exit()` or `std::exit()`, the
end-of-scope objects (e.g. from `lttng::make_scope_exit()` do not
fire.

Change-Id: I1ce6fd316356ac251cb3a0c0039c565f1ca1210f
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>