Fix: lttng-sessiond segfault during session destruction
In `cmd.c:cmd_destroy_session()`, if rotate_size is set for the
session, the condition is unsubscribed.

During unsubscription, the session's stored value is never nulled
out. It could happen that later, after the session is unlocked, the
rotate thread may try to do another round of unsubscribe/subscribe,
causing a double free of the condition.

Change-Id: I183a32224d25345a2088959e6d5c4e41a82d6801
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>