Clean-up: Tests: Provide default cwd when testing relayd working dir
In `tap.sh`, `is()` ends up aborting the test script if it is passed
an empty (null) value. When, for example, lttng-relayd fails to start,
the `cwd` will be empty as `readlink` will fail.

E.g. Output when a test fails and limits the runs:

```
...
ok 4 - lttng-relayd already started
  ---
    duration_ms: 12.914679
  ...

not ok 5 - Found lttng-relayd
  ---
    duration_ms: 49.119194
  ...

./tests/regression/tools/working-directory//../../..//utils/tap/tap.sh: line 288: 1: parameter null or not set
```

Change-Id: I54ac5c23a6d2604d261e8240679a827a12a57a1a
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>