misc: Allow users to override the location of clang-format
Since the coding imposes a specific version of clang-format (16),
the pre-commit hook can fail for reasons out of the contributor's
control.

Local git config settings are used to allow the user to specify
the location of these tools.

  git config --local hooks.clangFormatPath /path/to/clang-format-16

The settings are retrieved by the script and used if they were
specified.

Note that we should probably do the same for `black` as they
change the formatting from version to version. Right now I
feel we imply "use the latest version", but we should probably
impose one and update it on a release-to-release basis.

Change-Id: Iabb3e3964dc137075474ae4a240f30e0ecf4ec0d
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>