liblttng-ctl: Re-expose several `config_str_*` variables
Observed issue
==============

The diff of the ABI of liblttng-ctl between lttng-tools 2.13 and
master shows the following removed variables:

```

```
7 Removed variables:

  [D] 'const char* const config_str_false'    {config_str_false}
  [D] 'const char* const config_str_no'    {config_str_no}
  [D] 'const char* const config_str_off'    {config_str_off}
  [D] 'const char* const config_str_on'    {config_str_on}
  [D] 'const char* const config_str_true'    {config_str_true}
  [D] 'const char* const config_str_yes'    {config_str_yes}
  [D] 'log_time error_log_time'    {error_log_time}
```

Cause
=====

The `const_str_*` variables were moved from `session-config.cpp` to
`ini-config.cpp`. Due to license in-compatibility (and probably lack
of immediate need), libini-config was never linked to libconfig-lgpl,
and thus to liblttng-ctl.

Solution
========

Link to libini-config to libconfig.

Known drawbacks
===============

None.

Change-Id: I4fdb15157ca4a7d4ee34dff7d1dcaa44b2e23aef
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>