Tests: serial runner: fail test on abort (SIGINT and SIGTERM)
The CI kills the test suite using SIGTERM when it remains stuck
for a given time. When this occurs, it is hard to figure out which
test was running.

Catch both SIGTERM and SIGINT signals to gracefully report a test
failure that indicates the top-level test that was running when
the signal was received.

The number of planned tests is extracted to give a "hint" to the
user that the test could not complete.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: Id2aa025b43e9ddc205c784cc5e100fd04b46c3b9