tests: Retry taskset where possible
In the CI cluster, Incus will change the CPU affinities for containers
when another is started or stopped on the same host in an effort to
balance the work between all available CPUs on the host.

When this happens, the CPU chosen for the taskset may become
unavailable. These changes will detect the taskset failure and retry
as appropriate.

Change-Id: Idcaf1cecc155b618dd561a7d0f894ca19b3a5146
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>