tests: Force consumerd_testpoint to always be verbose
The `lttng_opt_*` variables aren't actually shared with the
consumerd. When running the consumerd with `--verbose`, the debug
messages from this library were not visible.

Using dlsym() to find the next `lttng_opt_verbose` does find a symbol,
but despite the consumerd running with `--verbose`, the value of that
symbol was 0.

As these testpoints are only used in a couple of situations during the
CI run, the local verbosity level can be forced high to always have
the information available.

Change-Id: Ife79be898f4534f16fc5ee587cb152bb1c591b31
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>