tests: Ignore notification type mismatches in `test_notification_ust_buffer_usage`
Observed Issue
==============

Frequent aborts in the CI on this test.

Cause
=====

The notification queue can be racy with regards the clients and
consumers being stopped and started as they are in this test. The
conditions for notifications are evaluated during each iteration of
the monitor thread, by default every ~ 1000000 us.

This means that in the moments between commands more than one
notification can be registered as the states of the buffers are
assessed.

Solution
========

Ignore notification with types that don't match the expected
notification type.

Known drawbacks
===============

None.

Change-Id: I62c269a218a473b53dc11809b99b7c45b0a9887a
Tested-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>