lttng-enable-channel: Add --buffer-ownership
The `--buffer-ownership' option accepts the following parameters:

  - user (equivalent to --buffers-uid)
  - process (equivalent to --buffers-pid)
  - system (equivalent to --buffers-global)

The parameter `system' is only valid with kernel domain.

The parameters `user' and `process' are only valid with userspace domain.

Ownership can be used multiple-times and can be used with
`--buffers-{uid,pid,global}' options.  The latest option having
precedence.

If no ownership is explicitly passed either with
`--buffers-{ownership,uid,pid,global}', then the ownership is defaulted
to:

  - `system' for kernel domain channel
  - `user' for userspace domain channel

Change-Id: I7749470a80950725f23d2fa61d86c7d77df235e2
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>