Tests: tools/channel: test --buffer-ownership and --buffers-* options
Add a test for legacy CLI buffering policies.

The test ensures that for a given session, creating two channels with
the new buffer policy option (--buffer-ownership={user,process,system})
and the legacy equivalent (--buffers-{uid,pid,global}) will yield
equivalent domain in the session and that both channels are in it.

Change-Id: I36cff84de1d56427a0c136be780aa4bb3316482f
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>