tests: Avoid failing on loaded modules when kernel tests are skipped
In the future I would like to be able to run lttng-sessiond of the
container host for various reasons. When loaded on the host as the
root user, the lttng-modules are loaded when available - and this is
visible inside the containers which share the kernel. Some tests
validate that the modules are unloaded even if the kernel testing has
been skipped.

The second part of this change limits the possible PID search space to
those started by the current user. This avoids situations where a
non-privileged user tries to kill the lttng-sessiond run by root.

Change-Id: I0cec2e385f37c62d8961488751805b1248ce2687
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>