Fix: build with urcu 0.15 on 32bit archs with 64bit time_t
In the health check code we use a 'struct timespec' which has a 'tv_sec'
member of type 'time_t'. On 32 bit architectures, this type has
traditionnaly been 32bit but is currently being transitioned to 64bit on
major distributions to handle y2k38.

We set this 64bit member with an urcu atomic operation which is
technically not supported on most 32bit architectures but would still
build with previous userspace-rcu versions. Starting with 0.15 these
atomic set operations are handled with compiler atomic builtins which
will use a support library with spinlocks when atomic operations
for a type are unavailable. This results in a build failure because
we don't link with this support library.

The good news is that we don't need to use atomic operations to
initialize the health state since this happens before it is added to the
global TLS state list.

Change-Id: I3fd3eab3a554ccd271ac961084b2966feef739c4
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>