sessiond: Only expect cpu_id for per-cpu channels
The session daemon emits the metadata for the packet header and expects
to have a 'cpu_id' field at the end of it.

However, this field is only emitted for per-cpu channels because the CPU
ID is essentially free due to the nature of the algorithm used for
selecting the stream to write to.

Therefore, this field is not emitted for other channel types.  This does
introduce a difference between traces generated with different buffer
allocation policies, even though the program instrumented was the same
and reproducible.

To overcome this, a new 'cpu_id' context ought to be added to support
the new buffer allocation policies that do not implicitly make the CPU
ID available.

Change-Id: I4344257b77889fe55288e754619c75986c249ee6
Depends-on: lttng-ust: I05107e08106b1dcd69dbde5bd76b9ff6ed9b15db
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>